initial spi-slave
_mostly_ works.  the "PRE" byte never comes out, and it hangs trying to
drain out the "final" byte, (FIFO_COUNT sticks at 1) but it's getting
the correct registers read out in teh same transaction, so that's
already about as well as we've ever done spi slave.

We added ETL here, as we think we'll probably end up  using it, and we
used in an early version of this code.