End of the road for tinyusb
WIP commit on where we went with tinyusb host, this was attempting to
update the tinyusb tree, but it's still borked even now. end of the
road.

note: right here this still runs on k64 and enumerates a mouse, so we
know the USB clock setup and IRQ is being handled properly.