saving our point in time.
This... should work? We've fixed our freertos config, we know the clocks
are right, and it enumerates the scanner and the 2/7 port hubs, but it
doesn't enumerate the mouse.  it works fine with the upstream example
though. let's build it up again from scratch, keeping as much of the
upstream code as we can